Particion del modelo en DataBase y Service
