+++ .gitignore
@@ -1260,3 +1260,27 @@ ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie
 ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/DataBase/room/ConceptDataBase_Impl.class
 ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/DataBase/room/ConceptDataBase_Impl$1.class
 ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/DataBase/room/ConceptDao.dex
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/ModelImp.dex
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/Model.dex
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/Service/Transform.class
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/Service/Service.class
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/Service/BighugelabsService.class
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/RepositoryImp.class
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/Service/BighugelabsService.dex
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/RepositoryImp.dex
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/view/BoldText.class
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/Service/WordsBighugelabsAPI.class
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/Service/TransformToString.class
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/view/BoldText.dex
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/Service/WordsBighugelabsAPI.dex
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/Repository.dex
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/ModelModule.dex
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/ModelListener.dex
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/Service/TransformToString.dex
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/Service/Transform.dex
+ADictionaryCharlie/app/build/intermediates/transforms/dexBuilder/debug/23/ayds/dictionary/charlie/fulllogic/model/Service/Service.dex
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/ModelImp.class
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/Model.class
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/Repository.class
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/ModelModule.class
+ADictionaryCharlie/app/build/intermediates/classes/debug/ayds/dictionary/charlie/fulllogic/model/ModelListener.class
