+++ ADictionaryCharlie/app/src/main/java/ayds/dictionary/charlie/fulllogic/controller/Controller.java
@@ -1,8 +1,5 @@
 package ayds.dictionary.charlie.fulllogic.controller;
 
-import ayds.dictionary.charlie.fulllogic.model.Service.Model;
-
 public interface Controller {
     void onEventUpdate(String searchedWord);
-    void setModel(Model model);
 }
