+++ ADictionaryCharlie/app/src/main/java/ayds/dictionary/charlie/fulllogic/controller/ControllerImp.java
@@ -1,19 +1,17 @@
 package ayds.dictionary.charlie.fulllogic.controller;
 
-import ayds.dictionary.charlie.fulllogic.model.Service.Model;
+import ayds.dictionary.charlie.fulllogic.model.Model;
 
 public class ControllerImp implements Controller {
 
     private Model model;
 
-    public ControllerImp() {}
+    public ControllerImp(Model model) {
+        this.model = model;
+    }
 
     @Override
     public void onEventUpdate(String searchedWord) {
         model.searchWord(searchedWord);
-    }
-
-    public void setModel(Model model){
-        this.model = model;
     }
 }
