+++ ADictionaryCharlie/app/src/main/java/ayds/dictionary/charlie/fulllogic/model/Model.java
@@ -1,0 +1,6 @@
+package ayds.dictionary.charlie.fulllogic.model;
+
+public interface Model {
+    void setListener(ModelListener listener);
+    void searchWord(String searchedWord);
+}
