+++ ADictionaryCharlie/app/src/main/java/ayds/dictionary/charlie/fulllogic/model/ModelListener.java
@@ -1,4 +1,4 @@
-package ayds.dictionary.charlie.fulllogic.model.Service;
+package ayds.dictionary.charlie.fulllogic.model;
 
 public interface ModelListener {
     void didUpdate(String lastSearch);
