--- ADictionaryCharlie/app/src/main/java/ayds/dictionary/charlie/fulllogic/model/Service/Model.java
@@ -1,9 +1,0 @@
-package ayds.dictionary.charlie.fulllogic.model.Service;
-
-import android.content.Context;
-
-public interface Model {
-    void createRepository(Context applicationContext);
-    void setListener(ModelListener listener);
-    void searchWord(String searchedWord);
-}
