+++ ADictionaryCharlie/app/src/main/java/ayds/dictionary/charlie/fulllogic/model/Service/Service.java
@@ -1,0 +1,5 @@
+package ayds.dictionary.charlie.fulllogic.model.Service;
+
+public interface Service {
+    String searchWord(String searchedWord);
+}
