+++ ADictionaryCharlie/app/src/main/java/ayds/dictionary/charlie/fulllogic/model/Service/Transform.java
@@ -1,0 +1,7 @@
+package ayds.dictionary.charlie.fulllogic.model.Service;
+
+import retrofit2.Response;
+
+public interface Transform {
+    String JSONtoString(Response<String> callResponse , String searchedWord);
+}
