+++ ADictionaryCharlie/app/src/main/java/ayds/dictionary/charlie/fulllogic/model/Service/WordsBighugelabsAPI.java
@@ -1,4 +1,4 @@
-package ayds.dictionary.charlie.fulllogic.model.DataBase;
+package ayds.dictionary.charlie.fulllogic.model.Service;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
