--- ADictionaryCharlie/app/src/main/java/ayds/dictionary/charlie/fulllogic/view/View.java
@@ -1,5 +1,0 @@
-package ayds.dictionary.charlie.fulllogic.view;
-
-public interface View {
-    void updateResult(String lastSearch);
-}
